Add blacklist check in say

+++ cogs/commands/fun/say.py
@@ -5,7 +5,7 @@ https://blacktooth-bot.com
 import discord
 
 from discord.ext import commands
-
+from cache import am as cache 
 class Say(commands.Cog):
     def __init__(self, client):
         self.client = client
@@ -15,6 +15,17 @@ class Say(commands.Cog):
         if 'https://' in message or 'http://' in message:
             await ctx.send("I'm afraid I can't send links.")
             return
+        if 'discord.gg' in message:
+            await ctx.send("I'm afraid I can't send invite links")
+            return 
+
+        if str(ctx.guild.id) in cache:
+            blacklist = cache[str(ctx.guild.id)]
+            for word in blacklist:
+                if word.lower() in message.lower():
+                    await ctx.send(f"Ayo {ctx.author.mention}, don't make me say blacklisted words!")
+                    return
+
         await ctx.message.delete()
         await ctx.send(f"{ctx.author.mention} said {message}", allowed_mentions=discord.AllowedMentions(
             everyone=False,
